batch: Move realized entry storage into package
The realization package owns entry views and provenance while their compact
mutable storage remains at the batch package root.

Content assembly needs all three parts of the same representation, and the
split location forces callers to cross a package boundary for storage
operations.

This commit moves realized-entry storage beneath realization and points it
at the nested view, provenance, and line-range modules.

Subsequent commits will move boundary lookup and mapped range construction
before realization coverage adopts only nested paths.
